♻️(backend) move playlist models to a dedicated module
Playlist will be used by all resources in Marsha. There is no more
reason to keep it in the video module so we move them in a dedicated
module called playlist.py
